Externalized component and add tests
